requirements.txt
fastapi==0.89.1
httpx==0.23.3
motor==3.1.1
Pillow==9.4.0
pydantic==1.10.4
pymongo==4.3.3
requests==2.28.2
starlette==0.22.0
uvicorn==0.20.0
websockets==10.4
python-dotenv==0.21.1
aioboto3==11.1.0